moved app commands to files in ecris.csd.viewer
